\name{otTensor-package}
\alias{otTensor-package}
\docType{package}
\title{
\packageTitle{otTensor}
}
\description{
\packageDescription{otTensor}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{otTensor}
\packageIndices{otTensor}
}
\author{
\packageAuthor{otTensor}

Maintainer: \packageMaintainer{otTensor}
}
\references{
Kerdoncuff, T. et al., (2022). Optimal Tensor Transport. \emph{Proceedings of the AAAI Conference on Artificial Intelligence}, 36(7), 7124-7132.
}

\keyword{ package }
\seealso{\code{\link{OTT}}}
\examples{
ls("package:otTensor")
}